refactor(apps/web): restructure player store and enhance playback actions
- Deleted legacy `player.store.ts` and `player.store.test.ts` files to streamline the codebase.
- Introduced new modular files for playback actions, queue actions, and server actions, improving organization and maintainability.
- Added comprehensive tests for playback and queue functionalities, ensuring robust state management and error handling.
- Enhanced synchronization logic with server state, improving playback consistency across devices.
